requirements.txt
# Django framework
Django==5.2.10

# Environment management
django-environ==0.11.2

# Database
psycopg2-binary==2.9.10

# Cache
redis==5.2.1

# WSGI server
gunicorn==23.0.0

# Image processing
pillow==11.0.0

# Utilities
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
packaging==24.2
asgiref==3.8.1

# Blog features
markdown==3.7
pygments==2.18.0

# Modern admin interface
django-unfold==0.40.0
